- Rename CartPage GraphQL files - Re-factor PDP configurable mutation to use cart fragment
