+++ packages/venia-ui/lib/components/CartPage/ProductListing/EditModal/productForm.js
@@ -9,7 +9,7 @@ import LoadingIndicator from '../../../LoadingIndicator';
 import Options from '../../../ProductOptions';
 import { QuantityFields } from '../quantity';
 import defaultClasses from './productForm.css';
-import { CartPageFragment } from '../../cartPageFragments';
+import { CartPageFragment } from '../../cartPageFragments.gql';
 
 const ProductForm = props => {
     const { item: cartItem, setIsCartUpdating } = props;
