+++ packages/venia-ui/lib/components/CartPage/cartPage.gql.js
@@ -1,5 +1,5 @@
 import gql from 'graphql-tag';
-import { CartPageFragment } from './cartPageFragments';
+import { CartPageFragment } from './cartPageFragments.gql';
 
 export const GET_CART_DETAILS = gql`
     query getCartDetails($cartId: String!) {
