+++ packages/venia-ui/lib/components/ProductFullDetail/productFullDetail.gql.js
@@ -1,0 +1,30 @@
+import gql from 'graphql-tag';
+
+import { CartPageFragment } from '../CartPage/cartPageFragments.gql';
+
+export const ADD_CONFIGURABLE_MUTATION = gql`
+    mutation addConfigurableProductToCart(
+        $cartId: String!
+        $quantity: Float!
+        $sku: String!
+        $parentSku: String!
+    ) {
+        addConfigurableProductsToCart(
+            input: {
+                cart_id: $cartId
+                cart_items: [
+                    {
+                        data: { quantity: $quantity, sku: $sku }
+                        parent_sku: $parentSku
+                    }
+                ]
+            }
+        ) {
+            cart {
+                id
+                ...CartPageFragment
+            }
+        }
+    }
+    ${CartPageFragment}
+`;
